pipeline/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.27.0
langgraph>=0.2.0
langchain-core>=0.3.0
google-generativeai>=0.8.0
pydantic>=2.0
